Major clean up - separate components into their own files
