+++ publisher/src/components/MetricConfiguration/index.ts
@@ -15,6 +15,9 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
+export * from "./Configuration";
+export * from "./ContextConfiguration";
+export * from "./MetricBox";
 export * from "./MetricConfiguration";
 export * from "./MetricConfiguration.styles";
 export * from "./MetricDefinitions";
